Added Galaga-like targets and Win/Loss Conditions
Now, you win when you survive 60 seconds, and lose if 5 targets get past you.

There are also rogue targets which go about twice as fast as the rest
of the pack.

I also modified the Hand's UI extensively, so that will need testing. I
also am unsure with how the "death"/"miss" counter is being represented.
If I could, I would make it little blue circles going away everytime you
miss a target. Sadly, I do not know how to do that.

Also, the Scene did not seem to push from the Lab, so a lot of the
components were not inserted. We'll have to test that I did the
reinsertion correctly.